refactor: move toArticlePage
